Move test to DomainTest.kt
