+++ src/test/kotlin/com/ojicoin/DomainTest.kt
@@ -1,20 +1,16 @@
 package com.ojicoin
 
-import com.ojicoin.domain.CreateViewCount
-import com.ojicoin.domain.ViewCounts
-import com.ojicoin.domain.toViewCount
+import com.ojicoin.domain.*
+import com.ojicoin.domain.Tags
 import com.ojicoin.service.DatabaseFactory
 import com.ojicoin.service.DatabaseFactory.dbQuery
 import kotlinx.coroutines.runBlocking
-import org.jetbrains.exposed.sql.SchemaUtils
-import org.jetbrains.exposed.sql.deleteAll
-import org.jetbrains.exposed.sql.insert
-import org.jetbrains.exposed.sql.selectAll
+import org.hamcrest.CoreMatchers
+import org.hamcrest.MatcherAssert
+import org.jetbrains.exposed.sql.*
+import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.transactions.transaction
-import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.RepeatedTest
+import org.junit.jupiter.api.*
 import kotlin.properties.Delegates
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
@@ -69,5 +65,44 @@ class DomainTest {
         internal fun tearDownAll() {
             transaction { SchemaUtils.drop(ViewCounts) }
         }
+    }
+}
+
+class UserTagTest {
+    @BeforeEach
+    internal fun setUp() {
+        DatabaseFactory.connectAndMigrate()
+        transaction {
+            SchemaUtils.create(UserTags)
+        }
+    }
+
+    @RepeatedTest(REPEATED_COUNT)
+    fun insert() = runBlocking {
+        // given
+        var userId by Delegates.notNull<Long>()
+        val createUserTag = fixture.giveMeOne(CreateUserTag::class.java)
+
+        // when
+        DatabaseFactory.dbQuery { userId = UserTags.insert { createUserTag.apply(it) } get UserTags.userId }
+
+        MatcherAssert.assertThat(userId, CoreMatchers.notNullValue())
+    }
+
+    @RepeatedTest(REPEATED_COUNT)
+    fun select() = runBlocking {
+        // given
+        lateinit var userTag: UserTag
+        var userId by Delegates.notNull<Long>()
+        val createUserTag = fixture.giveMeOne(CreateUserTag::class.java)
+        DatabaseFactory.dbQuery { userId = UserTags.insert { createUserTag.apply(it) } get UserTags.userId }
+
+        // when
+        DatabaseFactory.dbQuery {
+            userTag = UserTags.select { UserTags.userId eq userId }.firstNotNullOf { it.toUserTag() }
+        }
+
+        MatcherAssert.assertThat(userTag.tagId, CoreMatchers.equalTo(createUserTag.tagId))
+        MatcherAssert.assertThat(userTag.userId, CoreMatchers.equalTo(createUserTag.userId))
     }
 }
